user/web_api/serializers/otp_confirm_serializer.py
from rest_framework import serializers


class OtpConfirmSerializer(serializers.Serializer):
    email = serializers.EmailField()
    otp = serializers.IntegerField(max_value=999999, min_value=100000)