requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
distlib==0.3.8
Django==4.2.3
django-better-admin-arrayfield==1.4.2
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
django-js-asset==2.2.0
django-jsonfield-backport==1.0.5
django-postgres-extensions==0.9.3
django-postgres-extra==2.0.8
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
future==0.18.3
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.6
importlib-metadata==7.0.0
Markdown==3.5.1
packaging==23.2
Pillow==10.1.0
platformdirs==4.1.0
pluggy==1.3.0
promise==2.3
psycopg==3.1.15
psycopg-binary==3.1.13
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyproject-api==1.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
serializers==0.2.4
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
tox==4.11.4
typing_extensions==4.9.0
urllib3==2.1.0
virtualenv==20.25.0
zipp==3.17.0
drf-spectacular==0.27.0
django-cors-headers==4.3.1
djangorestframework-camel-case==1.4.2
DateTimeRange~=2.2.0
mbstrdecoder==1.1.3
typepy==1.3.2
jsonschema~=4.20.0